# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Link Shortener Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register user",
                "description": "register user ",
                "operationId": "88c674d9b8b26e4e27ccc2e547b37077",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "family",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "family": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "عملیات موفق"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "400"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "خطا در عملیات"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tokens/create": {
            "post": {
                "tags": [
                    "Tokens"
                ],
                "summary": "Create Token For Registered User",
                "description": "Create Token For Registered User ",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "api_key": {
                                                    "type": "string",
                                                    "default": "16|rYLBNvAAdtrXuGpM1WEQbIntRcC2iF5B2BD9SJsL"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "عملیات موفق"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "400"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "خطا در عملیات"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/link/shorten?link=your-link": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Shorten Link",
                "description": "Shorten Link",
                "operationId": "dfbd1b293724d3368656f370896f6068",
                "parameters": [
                    {
                        "name": "link",
                        "in": "query",
                        "description": "your link for shorten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "main-link": {
                                                    "type": "string",
                                                    "default": "https://www.zoomit.ir/tech-iran/384961-protection-of-childrens-online-data-neglected-in-iran/"
                                                },
                                                "shorten_link": {
                                                    "type": "string",
                                                    "default": "http://localhost:8000/EUhylqvfMI"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "عملیات موفق"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "400"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "خطا در عملیات"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "404"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "کاربر مورد نظر یافت نشد"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer ": []
                    }
                ]
            }
        },
        "/link/all": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Show All Links",
                "description": "Show All Links",
                "operationId": "e469ecf091e5ed90c5793ec32f7236e6",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "main-link": {
                                                    "type": "string",
                                                    "default": "https://www.zoomit.ir/tech-iran/384961-protection-of-childrens-online-data-neglected-in-iran/"
                                                },
                                                "shorten_link": {
                                                    "type": "string",
                                                    "default": "http://localhost:8000/EUhylqvfMI"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "default": "2022-08-12 11:12:25"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "عملیات موفق"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "400"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "خطا در عملیات"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "default": "404"
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "کاربر مورد نظر یافت نشد"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer ": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}